Move to pytorch inference (as it is specific to pytorch)
