+++ deeplabcut/pose_estimation_pytorch/modelzoo/__init__.py
@@ -16,3 +16,6 @@ from deeplabcut.pose_estimation_pytorch.modelzoo.utils import (
     get_super_animal_snapshot_path,
     load_super_animal_config,
 )
+from deeplabcut.pose_estimation_pytorch.modelzoo.inference import (
+    create_superanimal_inference_runners,
+)
