+++ deeplabcut/pose_estimation_pytorch/modelzoo/inference.py
@@ -12,7 +12,10 @@ import json
 from pathlib import Path
 
 import numpy as np
+import torch
 
+from deeplabcut.core.config import read_config_as_dict
+from deeplabcut.modelzoo.weight_initialization import build_weight_init
 from deeplabcut.modelzoo.utils import get_super_animal_scorer, get_superanimal_colormaps
 from deeplabcut.pose_estimation_pytorch.apis.videos import (
     create_df_from_prediction,
@@ -26,7 +29,10 @@ from deeplabcut.pose_estimation_pytorch.apis.utils import (
 )
 from deeplabcut.pose_estimation_pytorch.modelzoo.utils import (
     raise_warning_if_called_directly,
+    load_super_animal_config,
+    update_config,
 )
+from deeplabcut.pose_estimation_pytorch.runners import InferenceRunner
 from deeplabcut.utils.make_labeled_video import create_video
 
 
@@ -219,3 +225,128 @@ def _video_inference_superanimal(
             print(f"Video with predictions was saved as {output_video}")
 
     return results
+
+
+def create_superanimal_inference_runners(
+    superanimal_name: str,
+    model_name: str,
+    detector_name: str | None = None,
+    max_individuals: int = 10,
+    batch_size: int = 1,
+    detector_batch_size: int = 1,
+    device: str | None = "auto",
+    customized_model_config: str | Path | dict | None = None,
+    customized_pose_checkpoint: str | Path | None = None,
+    customized_detector_checkpoint: str | Path | None = None,
+) -> tuple[InferenceRunner, InferenceRunner, dict]:
+    """Create SuperAnimal inference runners for in-memory batched inference.
+
+    This helper is intended for Model Zoo inference pipelines that run directly on
+    arrays. It prepares pose/detector runners and returns them with the resolved
+    model config.
+
+    Args:
+        superanimal_name: Name of the SuperAnimal dataset, e.g.
+            ``"superanimal_quadruped"``.
+        model_name: Pose model architecture name, e.g. ``"hrnet_w32"``.
+        detector_name: Detector architecture name for top-down inference, e.g.
+            ``"fasterrcnn_resnet50_fpn_v2"``.
+        max_individuals: Maximum number of individuals to keep per frame.
+        batch_size: Batch size for pose inference.
+        detector_batch_size: Batch size for detector inference.
+        device: Device for inference. If ``"auto"`` or ``None``, resolves to CUDA
+            when available, else CPU.
+        customized_model_config: Optional path or dict for a custom model config.
+            If not provided, uses the default SuperAnimal config.
+        customized_pose_checkpoint: Optional custom pose checkpoint path.
+        customized_detector_checkpoint: Optional custom detector checkpoint path.
+
+    Returns:
+        tuple: ``(pose_runner, detector_runner, model_cfg)`` where:
+            - ``pose_runner`` is the pose inference runner
+            - ``detector_runner`` is the detector inference runner
+            - ``model_cfg`` is the resolved model configuration dict
+
+    Example:
+        >>> from pathlib import Path
+        >>> import numpy as np
+        >>> from PIL import Image
+        >>> from deeplabcut.pose_estimation_pytorch.modelzoo.inference import (
+        ...     create_superanimal_inference_runners,
+        ... )
+        >>>
+        >>> img_paths = [
+        ...     "/path/to/images/frame_0000.png",
+        ...     "/path/to/images/frame_0001.png",
+        ...     "/path/to/images/frame_0002.png",
+        ... ]
+        >>> images = [np.asarray(Image.open(Path(p)).convert("RGB")) for p in img_paths]
+        >>>
+        >>> pose_runner, det_runner, model_cfg = create_superanimal_inference_runners(
+        ...     superanimal_name="superanimal_quadruped",
+        ...     model_name="hrnet_w32",
+        ...     detector_name="fasterrcnn_resnet50_fpn_v2",
+        ...     max_individuals=10,
+        ...     batch_size=1,
+        ...     detector_batch_size=1,
+        ... )
+        >>>
+        >>> det_preds = det_runner.inference(images)
+        >>> pose_inputs = list(zip(images, det_preds))
+        >>> pose_preds = pose_runner.inference(pose_inputs)
+        >>> print(len(det_preds), len(pose_preds))
+    """
+    if model_name.startswith("fmpose3d"):
+        raise NotImplementedError(
+            "FMPose3D is not supported in this helper. Use the FMPose3D inference API."
+        )
+
+    if superanimal_name == "superanimal_humanbody":
+        raise NotImplementedError(
+            "superanimal_humanbody is currently not supported by this helper because "
+            "it relies on modelzoo.build_weight_init, which does not support this dataset."
+        )
+
+    if detector_name is None:
+        raise ValueError(
+            "Please provide `detector_name` for SuperAnimal top-down inference setup."
+        )
+
+    if device in (None, "auto"):
+        device = "cuda:0" if torch.cuda.is_available() else "cpu"
+
+    if customized_model_config is not None:
+        if isinstance(customized_model_config, (str, Path)):
+            model_cfg = read_config_as_dict(customized_model_config)
+        else:
+            model_cfg = customized_model_config.copy()
+    else:
+        model_cfg = load_super_animal_config(
+            super_animal=superanimal_name,
+            model_name=model_name,
+            detector_name=detector_name,
+        )
+
+    model_cfg = update_config(model_cfg, max_individuals=max_individuals, device=device)
+    weight_init = build_weight_init(
+        cfg=model_cfg,
+        super_animal=superanimal_name,
+        model_name=model_name,
+        detector_name=detector_name,
+        with_decoder=False,
+        memory_replay=False,
+        customized_pose_checkpoint=customized_pose_checkpoint,
+        customized_detector_checkpoint=customized_detector_checkpoint,
+    )
+
+    pose_runner, detector_runner = get_inference_runners(
+        model_config=model_cfg,
+        snapshot_path=weight_init.snapshot_path,
+        max_individuals=max_individuals,
+        num_bodyparts=len(model_cfg["metadata"]["bodyparts"]),
+        num_unique_bodyparts=len(model_cfg["metadata"]["unique_bodyparts"]),
+        batch_size=batch_size,
+        detector_batch_size=detector_batch_size,
+        detector_path=weight_init.detector_snapshot_path,
+    )
+    return pose_runner, detector_runner, model_cfg
